Split into abstract class
