+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Container/CompilerPasses/AbstractConfigureSchemaNamespacingCompilerPass.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GraphQLAPI\GraphQLAPI\Container\CompilerPasses;
+
+use PoP\ComponentModel\Schema\SchemaNamespacingServiceInterface;
+use PoP\Root\Container\CompilerPasses\AbstractCompilerPass;
+use PoP\Root\Container\ContainerBuilderWrapperInterface;
+
+abstract class AbstractConfigureSchemaNamespacingCompilerPass extends AbstractCompilerPass
+{
+    protected function doProcess(ContainerBuilderWrapperInterface $containerBuilderWrapper): void
+    {
+        $schemaNamespacingServiceDefinition = $containerBuilderWrapper->getDefinition(SchemaNamespacingServiceInterface::class);
+        $schemaNamespace = $this->getSchemaNamespace();
+        foreach ($this->getComponentClasses() as $componentClass) {
+            $componentClassNamespace = substr($componentClass, 0, strrpos($componentClass, '\\'));
+            $schemaNamespacingServiceDefinition->addMethodCall(
+                'addSchemaNamespaceForClassOwnerAndProjectNamespace',
+                [
+                    $componentClassNamespace,
+                    $schemaNamespace
+                ]
+            );
+        }
+    }
+
+    /**
+     * @return string[]
+     */
+    abstract protected function getComponentClasses(): array;
+    abstract protected function getSchemaNamespace(): string;
+}
